Unify header and clean up
