+++ src/components/StoreLayout.js
@@ -1,42 +1,22 @@
 import React from "react"
 import PropTypes from "prop-types"
-import { useStaticQuery, graphql, Link } from "gatsby"
 
-import Cart from "./Cart"
+import Header from "./Header"
 import ProductsProvider from "./ProductsProvider"
 import CartProvider from "./CartProvider"
 import IdentityProvider from "./IdentityProvider"
 import css from "./StoreLayout.module.css"
 
 const StoreLayout = ({ children }) => {
-  const data = useStaticQuery(graphql`
-    query SiteTitleQuery {
-      site {
-        siteMetadata {
-          title
-        }
-      }
-    }
-  `)
   return (
     <IdentityProvider>
       <ProductsProvider>
         <CartProvider>
-          <header className={css.header}>
-            <nav>
-              <Link to="/admin">← Admin</Link>
-              <h1 className={css.title}>
-                <Link to="/">{data.site.siteMetadata.title}</Link>
-              </h1>
-              <div>
-                <Cart />
-              </div>
-            </nav>
-          </header>
+          <Header cart></Header>
 
           <main className={css.main}>{children}</main>
 
-          <footer style={{ textAlign: "center", margin: "3rem 0" }}>
+          <footer style={{ textAlign: "center", margin: "2rem 0" }}>
             <span>
               Fork
               {` `}
