+++ src/components/admin/Layout.js
@@ -1,19 +1,13 @@
 import React from "react"
 import PropTypes from "prop-types"
-import { Link } from "gatsby"
 
 import "../hiq.css"
+import Header from "../Header"
 import ProductsProvider from "./ProductsProvider"
-import css from "./Layout.module.css"
 
 const Layout = ({ children }) => (
   <ProductsProvider>
-    <header className={css.header}>
-      <nav className={css.nav}>
-        <Link to="/admin">Admin</Link>
-        <Link to="/">Store →</Link>
-      </nav>
-    </header>
+    <Header></Header>
     <main>{children}</main>
   </ProductsProvider>
 )
