--- src/components/admin/Layout.module.css
@@ -1,12 +1,0 @@
-.header {
-  border-bottom: 1px solid #ddd;
-  margin-bottom: 0.5rem;
-}
-
-.nav {
-  display: flex;
-  justify-content: space-between;
-  margin: 0 auto;
-  max-width: 1000px;
-  padding: 1rem 0;
-}
