+++ src/components/admin/ProductForm.module.css
@@ -2,7 +2,6 @@
   background-color: white;
   border-bottom: 1px solid #ddd;
   padding: 0.5rem;
-  margin-top: -0.5rem;
   position: sticky;
   top: 0;
   width: 100%;
